Rename timeline view modules
